blog_project/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
apptools==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bleach==4.1.0
blinker==1.4
bottle==0.12.19
bs4==0.0.1
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.1.3
cmake==3.25.0
colorama==0.4.4
configobj==5.0.6
coverage==6.2
cycler==0.10.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.6.0
Django==3.2.11
django-appconf==1.0.5
django-bootstrap4==22.1
django-compressor==3.1
django-crispy-forms==1.14.0
django-haystack==3.1.1
django-ipware==4.0.2
django-markdown-deux==1.0.5
django-mdeditor==0.1.18
django-simpleui==2022.4.9
django-uuslug==1.2.0
easygui==0.98.3
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
entrypoints==0.4
envisage==6.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.6.0
Flask==2.1.2
Flask-Mail==0.9.1
flatbuffers==22.10.26
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
gevent==21.12.0
git-review==2.2.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.50.0
h5py==3.7.0
idna==2.10
importlib-metadata==4.10.1
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
joblib==0.16.0
jsonpickle==2.0.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
libclang==14.0.6
lxml==4.9.1
Markdown==3.3.6
markdown2==2.4.3
MarkupSafe==2.1.1
matplotlib==3.3.2
matplotlib-inline==0.1.6
mayavi==4.8.1
mistune==2.0.4
mplfinance==0.12.9b7
multidict==6.0.2
mysqlclient==2.1.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyface==7.4.3
Pygments==2.10.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==2.12.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.2
PySide2==5.15.2.1
PySide6==6.4.1
PySide6-Addons==6.4.1
PySide6-Essentials==6.4.1
python-dateutil==2.8.2
python-logstash==0.4.6
python-memcached==1.59
python-slugify==5.0.2
pytz==2021.3
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.9
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
raven==6.10.0
rcssmin==1.1.0
requests==2.26.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
rsa==4.7.2
scikit-build==0.12.0
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.8.0
shiboken2==5.15.2.1
shiboken6==6.4.1
simplejson==3.18.0
six==1.15.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.6.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.17.0
text-unidecode==1.3
threadpoolctl==2.1.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.5.0
traits==6.4.1
traitsui==7.4.2
tushare==1.2.89
typing_extensions==4.4.0
tzdata==2021.5
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.2.0
virtualenv==20.14.1
vtk==9.2.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.1.2
WeRoBot==1.13.1
Whoosh==2.7.4
widgetsnbextension==4.0.3
wrapt==1.14.1
wslink==1.9.1
xlrd==2.0.1
xlrd2==1.3.4
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.8.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0